requirements/local.txt
-r base.txt
psycopg2-binary==2.8.6
django-debug-toolbar==3.2.2
